% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucide.R
\name{lucide_list}
\alias{lucide_list}
\title{List available Lucide icons}
\usage{
lucide_list()
}
\value{
A character vector of icon names.
}
\description{
Returns the names of all icons bundled in this version of 'lucidr'.
}
\examples{
head(lucide_list(), 10)
length(lucide_list())

}
